Automatically run EdgeDB queries in a transaction for GQL mutations
