+++ src/core/edgedb/edgedb.module.ts
@@ -1,8 +1,10 @@
 import { Module, OnModuleDestroy } from '@nestjs/common';
+import { APP_INTERCEPTOR } from '@nestjs/core';
 import { createClient } from 'edgedb';
 import { KNOWN_TYPENAMES } from 'edgedb/dist/codecs/consts.js';
 import { ScalarCodec } from 'edgedb/dist/codecs/ifaces.js';
 import { Class } from 'type-fest';
+import { EdgeDBTransactionalMutationsInterceptor } from './edgedb-transactional-mutations.interceptor';
 import { EdgeDB } from './edgedb.service';
 import { Client } from './reexports';
 import { LuxonCalendarDateCodec, LuxonDateTimeCodec } from './temporal.codecs';
@@ -25,6 +27,10 @@ import { TransactionContext } from './transaction.context';
     },
     EdgeDB,
     TransactionContext,
+    {
+      provide: APP_INTERCEPTOR,
+      useClass: EdgeDBTransactionalMutationsInterceptor,
+    },
   ],
   exports: [EdgeDB, Client],
 })
